refactor(settings): extract RulesPage
The last and biggest page extraction (990 lines): multi-step rule
conditions/actions, priority reordering, live validation feedback, and
the rule-test dialog hookup. Moved wholesale into gui/settings/rules.py
per the plan -- no logic changes, only hoisting repeated local imports
(get_theme_manager, WheelIgnoreComboBox, QComboBox/QLineEdit/QDateEdit,
etc.) to module level, matching the pattern already used by the other
extracted pages.

save_settings()'s inline Rules block is gone; RulesPage.collect() owns
it now, going through the shell's ordinary validate/collect loop.
